Rename the SimpleSubmitPinComponent to SubmitPinFormComponent
